+++ yggdrasil_decision_forests/port/python/CHANGELOG.md
@@ -19,6 +19,8 @@
 -   Expose the slow engine for debugging predictions and evaluations with 
     `use_slow_engine=True`.
 -   Speed-up training of GBT models by ~10%
+-   Add `ydf.util.read_tf_record` and `ydf.util.write_tf_record` to facilitate
+    TF Record datasets usage.
 
 ## 0.7.0 - 2024-08-21